Converted back to using double as decimal has much smaller min and max values

+++ Calculator/Form1.cs
@@ -21,14 +21,14 @@ namespace Calculator
          *      - Use enum to make CE work properly when '=' was last operation.
          * Unit Tests?
          * Can we have the number seperately from the textbox, then update the textbox whenever something changes 
-         *  - So we aren't converting back and forth from decimal and string
+         *  - So we aren't converting back and forth from double and string
          */
 
 
 
 
 
-        decimal remembered_characters;
+        double remembered_characters;
         
         public Calculator()
         {
@@ -51,7 +51,7 @@ namespace Calculator
 
 
 
-        private void add_character(decimal character)
+        private void add_character(double character)
         {
             if (textBoxBottom.Text.Length < 18) {
                 if (operation_enum != Operation_Enum.Equals) {
@@ -74,7 +74,7 @@ namespace Calculator
         {
             if (textBoxBottom.Text.Length != 0)
             {
-                remembered_characters = Convert.ToDecimal(textBoxBottom.Text);
+                remembered_characters = Convert.ToDouble(textBoxBottom.Text);
                 textBoxBottom.Text = "";
             }
         }
@@ -168,17 +168,17 @@ namespace Calculator
         }
         private void button_square_Click(object sender, EventArgs e)
         {
-            textBoxBottom.Text = Math.Pow((double)decimal.Parse(textBoxBottom.Text), 2).ToString();
+            textBoxBottom.Text = Math.Pow((double)double.Parse(textBoxBottom.Text), 2).ToString();
         }
 
         private void button_square_root_Click(object sender, EventArgs e)
         {
-            textBoxBottom.Text = Math.Sqrt((double)decimal.Parse(textBoxBottom.Text)).ToString();
+            textBoxBottom.Text = Math.Sqrt((double)double.Parse(textBoxBottom.Text)).ToString();
         }
 
         private void button_equal_Click(object sender, EventArgs e)
         {
-            decimal ToOperate = decimal.Parse(textBoxBottom.Text);
+            double ToOperate = double.Parse(textBoxBottom.Text);
 
             textBoxTop.Text += textBoxBottom.Text + " = ";
 
@@ -237,7 +237,7 @@ namespace Calculator
         {
             if (remembered_characters != 0)
             {
-                textBoxBottom.Text = (remembered_characters * (decimal.Parse(textBoxBottom.Text)/100)).ToString();
+                textBoxBottom.Text = (remembered_characters * (double.Parse(textBoxBottom.Text)/100)).ToString();
             } 
             else
             {
@@ -247,8 +247,8 @@ namespace Calculator
 
         private void button_One_Over_Click(object sender, EventArgs e)
         {
-            if (decimal.Parse(textBoxBottom.Text) != 0) {
-                textBoxBottom.Text = (1 / decimal.Parse(textBoxBottom.Text)).ToString();
+            if (double.Parse(textBoxBottom.Text) != 0) {
+                textBoxBottom.Text = (1 / double.Parse(textBoxBottom.Text)).ToString();
             } else
             {
                 textBoxBottom.Text = "Cannot divide by zero";
